Made the unit tests more descriptive

+++ src/main/java/no/stelar7/api/l4j/dto/summoner/Summoner.java
@@ -34,7 +34,7 @@ public class Summoner implements Serializable
     String name;
     int    profileIconId;
     Long   revisionDate;
-    Long   summonerLevel;
+    int   summonerLevel;
     Server region;
 
     public List<League> getFullLeague() throws LibraryException
